save n as integer and p as float'

+++ math/0x03-probability/binomial.py
@@ -30,7 +30,7 @@ class Binomial:
             if not 0 < p < 1:
                 raise ValueError('p must be greater than 0 and less than 1')
             else:
-                self.n = float(n)
+                self.n = int(n)
                 self.p = float(p)
         else:
             if not isinstance(data, list):
